Harden finance service for legacy columns in KPI, balance and cashflow
